// src/app/interface/tiempo.liturgia.interface.ts
// Generated by https://quicktype.io

export interface ResultTiempoLiturgias {
  ok:             boolean;
  msg:            string;
  tiempoliturgia: Tiempoliturgia[];
}

export interface ResultTiempoLiturgia {
  ok:             boolean;
  msg:            string;
  tiempoliturgia: Tiempoliturgia;
}

export interface Tiempoliturgia {
  id:          number;
  descripcion: string;
  estado:      number;
}

export interface TiempoLiturgiaForm{
  descripcion:string
}