// src/app/interface/tipo.liturgia.interface.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface ResultTipoLiturgias {
  ok:           boolean;
  msg:          string;
  tipoliturgia: Tipoliturgia[];
}

export interface ResultTipoLiturgia{
  ok:           boolean;
  msg:          string;
  tipoliturgia: Tipoliturgia;
}

export interface Tipoliturgia {
  id:              number;
  fecha:           string;
  nom_fecha:       string;
  descripcion:     string;
  id_liturgia:     number;
  id_tiempo:       number;
  TiempoLiturgium: TiempoLiturgium;
  Liturgium:       Liturgium;
}

export interface Liturgium {
  id:     number;
  nombre: string;
  logo:   string;
  estado: number;
}

export interface TiempoLiturgium {
  id:          number;
  descripcion: string;
  estado:      number;
}




export interface TipoLiturgiaForm{
  fecha:string,
  nom_fecha:string,
  descripcion:string,
  id_liturgia:string,
  id_tiempo:string
}